Delete Hank Chores
