+++ src/features/world/ui/PlayerModals.tsx
@@ -19,9 +19,9 @@ import { Button } from "components/ui/Button";
 import { Revealed } from "features/game/components/Revealed";
 import { ChestRevealing } from "./chests/ChestRevealing";
 import { secondsToString } from "lib/utils/time";
-import { secondsTillReset } from "features/helios/components/hayseedHank/HayseedHankV2";
 import { AdminSettings } from "features/island/hud/components/settings-menu/general-settings/AdminSettings";
 import { CONFIG } from "lib/config";
+import { secondsTillReset } from "lib/utils/time";
 
 type Player = {
   id: number;
